few changes

+++ BlinkyTest/src/Modules/GPSModule.h
@@ -7,11 +7,10 @@ class GPSModule: public SoftwareSerial
 {
   public:
      GPSModule(SDCardModule* sd, uint8_t rxPin, uint8_t txPin);
-     bool process(uint8_t count);
+     bool process(uint32_t count);
      void begin();
 
   private:
     static const uint32_t BAUDRATE = 9600;
     SDCardModule* sd;
-
 };
